restructured folders, added additional keys
